+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.0.2 - 2025-09-19
+- Add initial project files including HTML, CSS, JS, and service worker for PWA functionality
+- Adds web components for banners, toasts, network status, install and launch app buttons
 ## 0.0.1 - 2025-09-19
 - Initialize repository and base project setup.
 - Add README file with minimal description